Feature/usermanagement

+++ app/User.php
@@ -41,14 +41,77 @@ class User extends Authenticatable
         }
     }
 
+    public function canModerate()
+    {
+        if ($this->isAdmin() || $this->isModerator()) {
+            return true;
+        }
+
+        return false;
+    }
+
+    public function home()
+    {
+        if ($this->canModerate()) {
+            return route('admin.post.index');
+        }
+
+        return '/admin';
+    }
+
+    public function isAdmin()
+    {
+        return strtolower($this->level()) == 'admin';
+    }
+
+    public function isModerator()
+    {
+        return strtolower($this->level()) == 'moderator';
+    }
+
+    public function metas()
+    {
+        return $this->belongsToMany('App\Meta')->withPivot('value')->withTimestamps();
+    }
+
+    public function level()
+    {
+        return $this->meta('level') ?? 'user';   
+    }
+
+    public function meta($code = null)
+    {
+        if (empty($code)) {
+            return null;
+        }
+
+        if ($meta = $this->metas()->where('code', $code)->first()) {
+            return $meta->pivot->value;
+        }
+
+        return null;
+    }
+
+    public function updateMeta($code = null, $value = null)
+    {
+        if (!empty($code)) {
+            return $this->metas()->sync([Meta::where('code', $code)->first()->id => ['value' => $value] ], false);
+        }
+
+        return false;
+    }
+
     public static function routes()
     {
-        Route::get('/admin/user/', 'Admin\UserController@index')->name('admin.user.index');
-        Route::get('/admin/user/{user}', 'Admin\UserController@show')->name('admin.user.show');
-        Route::get('/admin/user/create', 'Admin\UserController@create')->name('admin.user.create');
-        Route::post('/admin/user/store', 'Admin\UserController@store')->name('admin.user.store');
-        Route::get('/admin/user/{user}/edit', 'Admin\UserController@edit')->name('admin.user.edit');
-        Route::patch('/admin/user/{user}', 'Admin\UserController@update')->name('admin.user.update');
-        Route::delete('/admin/user/{user}', 'Admin\UserController@destroy')->name('admin.user.destroy');
+        Route::get('/admin/user/', 'Admin\UserController@index')->name('admin.user.index')->middleware(['auth','moderator']);
+        Route::get('/admin/user/{user}/edit', 'Admin\UserController@edit')->name('admin.user.edit')->middleware(['auth','moderator']);
+        Route::patch('/admin/user/{user}', 'Admin\UserController@update')->name('admin.user.update')->middleware(['auth','moderator']);
+        Route::delete('/admin/user/{user}', 'Admin\UserController@destroy')->name('admin.user.destroy')->middleware(['auth','moderator']);
+
+        Route::get('/user/profile/{user}', 'Profile\UserController@show')->name('profile.show')->middleware('auth');
+        Route::get('/user/profile/{user}/edit', 'Profile\UserController@edit')->name('profile.edit')->middleware('auth');
+        Route::patch('/user/profile/{user}', 'Profile\UserController@update')->name('profile.update')->middleware('auth');
+        // Route::get('/', 'Admin\PostController@index')->name('profile.user.show');
+        // Route::get('/home', 'Admin\PostController@index')->name('home');
     }
 }
